Improve implementation, split up into separate files
