+++ .gitignore
@@ -6,4 +6,4 @@ __pycache__
 *.pkl
 *.egg-info
 MUJOCO_LOG.TXT
-hydrax/benchmark_results/
+hydrax/benchmark/results/
